Adapt SurveyUrlProvider.kt so that now the otp has to be passed directly
