+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/datadonation/survey/SurveyUrlProviderTest.kt
@@ -1,0 +1,53 @@
+package de.rki.coronawarnapp.datadonation.survey
+
+import de.rki.coronawarnapp.appconfig.AppConfigProvider
+import io.kotest.assertions.throwables.shouldThrow
+import io.kotest.matchers.shouldBe
+import io.mockk.MockKAnnotations
+import io.mockk.clearAllMocks
+import io.mockk.coEvery
+import io.mockk.impl.annotations.MockK
+import kotlinx.coroutines.test.runBlockingTest
+import okhttp3.HttpUrl.Companion.toHttpUrl
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import java.util.UUID
+
+internal class SurveyUrlProviderTest {
+
+    @MockK lateinit var appConfigProvider: AppConfigProvider
+
+    @BeforeEach
+    fun setUp() {
+        MockKAnnotations.init(this)
+        coEvery { appConfigProvider.getAppConfig().survey.otpQueryParameterName } returns "queryParamNameOtp"
+    }
+
+    @AfterEach
+    fun tearDown() {
+        clearAllMocks()
+    }
+
+    private fun createInstance() = SurveyUrlProvider(
+        appConfigProvider,
+    )
+
+    @Test
+    fun `provideUrl() should provide correct Url`() = runBlockingTest {
+        val otp = UUID.randomUUID()
+        coEvery { appConfigProvider.getAppConfig().survey.surveyOnHighRiskUrl } returns "http://www.example.com".toHttpUrl()
+
+        createInstance().provideUrl(Surveys.Type.HIGH_RISK_ENCOUNTER, otp) shouldBe "http://www.example.com/?queryParamNameOtp=$otp"
+    }
+
+    @Test
+    fun `provideUrl() should throw IllegalStateException when url from AppConfig is null`() = runBlockingTest {
+        val otp = UUID.randomUUID()
+        coEvery { appConfigProvider.getAppConfig().survey.surveyOnHighRiskUrl } returns null
+
+        shouldThrow<IllegalStateException> { createInstance().provideUrl(Surveys.Type.HIGH_RISK_ENCOUNTER, otp) }.also {
+            it.message shouldBe "AppConfig doesn't contain a link to the high-risk card survey"
+        }
+    }
+}
